WIP: Asynchronously install missing devdocs
If docs were configured but aren't installed, they will be scheduled
for installation during idle time the first time a buffer for the
corresponsing major mode is opened. This makes the docs available
without user intervention, but doesn't block editing when the buffer
is first opened.

Currently, there are weird bugs in `run-with-idle-timer` that make
rescheduling unreliable if the user starts typing and interrupts the
installation.